+++ src/main/java/org/eclipselabs/garbagecat/domain/jdk/CmsInitialMarkEvent.java
@@ -159,7 +159,7 @@ public class CmsInitialMarkEvent extends CmsCollector implements BlockingEvent, 
         return timeReal;
     }
 
-    public byte getParallelism() {
+    public int getParallelism() {
         return JdkMath.calcParallelism(timeUser, timeReal);
     }
