+++ src/main/java/org/eclipselabs/garbagecat/domain/jdk/G1YoungPauseEvent.java
@@ -320,7 +320,7 @@ public class G1YoungPauseEvent extends G1Collector
         return timeReal;
     }
 
-    public byte getParallelism() {
+    public int getParallelism() {
         return JdkMath.calcParallelism(timeUser, timeReal);
     }
